+++ text-repository/src/main/java/eu/interedition/text/repository/TokenizerSettings.java
@@ -9,11 +9,11 @@ import java.util.Set;
  * @author <a href="http://gregor.middell.net/" title="Homepage">Gregor Middell</a>
  */
 public interface TokenizerSettings {
-  boolean startingAnnotationsAreBoundary(Text text, int offset, Set<Annotation> annotations);
+  boolean startingAnnotationsAreBoundary(Text text, long offset, Set<Annotation> annotations);
 
-  boolean emptyAnnotationsAreBoundary(Text text, int offset, Set<Annotation> annotations);
+  boolean emptyAnnotationsAreBoundary(Text text, long offset, Set<Annotation> annotations);
 
-  boolean endingAnnotationsAreBoundary(Text text, int offset, Set<Annotation> annotations);
+  boolean endingAnnotationsAreBoundary(Text text, long offset, Set<Annotation> annotations);
 
-  boolean isBoundary(Text text, int offset, char c);
+  boolean isBoundary(Text text, long offset, char c);
 }
